move React logic out of exclusion zone Open Layers, put in Map.tsx where it belongs
