+++ src/web/openlayers/layers/vector/exclusion-zone-layer.ts
@@ -1,18 +1,16 @@
 import { Feature } from "ol";
 import { Point, Polygon } from "ol/geom";
 import { Draw } from "ol/interaction";
-import { DrawEvent } from "ol/interaction/Draw";
 import Fill from "ol/style/Fill";
 import Stroke from "ol/style/Stroke";
 import Style from "ol/style/Style";
 import CircleStyle from "ol/style/Circle";
 import Text from "ol/style/Text";
-import { fromLonLat, toLonLat } from "ol/proj";
+import { fromLonLat } from "ol/proj";
 
 import JaiaVectorLayer from "./jaia-vector-layer";
 import { layersZIndexes } from "../zindex";
 import { LayerTitles, MapFeatureTypes } from "../../../types/openlayers-types";
-import { JaiaActions } from "../../../context/jaia-actions";
 import { getZoneBufferVertices } from "../../../data/exclusion_zones/exclusion-zone-router";
 import { jaiaGlobal } from "../../../data/jaia_global/jaia-global";
 import { exclusionZoneSet } from "../../../data/exclusion_zones/exclusion-zone-set";
@@ -28,8 +26,6 @@ const BUFFER_FILL = "rgba(255, 140, 0, 0.06)";
 
 class ExclusionZoneLayer extends JaiaVectorLayer {
     private draw: Draw | null = null;
-    private dispatch: ((action: { type: JaiaActions; [key: string]: unknown }) => void) | null =
-        null;
 
     constructor() {
         super(
@@ -37,16 +33,6 @@ class ExclusionZoneLayer extends JaiaVectorLayer {
             layersZIndexes.get(LayerTitles.EXCLUSION_ZONE_LAYER),
         );
         this.getVectorLayer().setStyle(this.getZoneStyle.bind(this));
-    }
-
-    /**
-     * Wires up the React dispatch so the draw interaction can fire actions
-     *
-     * @param {Function} dispatch The dispatch function from JaiaContext
-     * @returns {void}
-     */
-    setDispatch(dispatch: (action: { type: JaiaActions; [key: string]: unknown }) => void) {
-        this.dispatch = dispatch;
     }
 
     /**
@@ -59,8 +45,9 @@ class ExclusionZoneLayer extends JaiaVectorLayer {
     }
 
     /**
-     * Creates a polygon Draw interaction. When the operator finishes drawing,
-     * dispatches ADD_EXCLUSION_ZONE with the polygon vertices in lat/lon.
+     * Creates and stores a polygon Draw interaction configured with the zone drawing style.
+     * The drawend event must be wired by the caller (Map.tsx) so dispatch stays inside
+     * the React boundary and is never stored on this singleton.
      *
      * @returns {Draw} The configured Draw interaction
      */
@@ -73,42 +60,7 @@ class ExclusionZoneLayer extends JaiaVectorLayer {
                 stroke: new Stroke({ color: DRAW_STROKE, width: 2, lineDash: [6, 4] }),
             }),
         });
-
-        this.draw.on("drawend", (event: DrawEvent) => {
-            const feature = event.feature as Feature<Polygon>;
-            const coords3857 = feature.getGeometry().getCoordinates()[0];
-
-            // OpenLayers closes the ring by repeating the first vertex — skip it.
-            // Also deduplicate consecutive identical vertices (double-click to finish
-            // registers the last vertex twice, causing false non-convex detection).
-            const allVertices = coords3857.slice(0, -1).map((coord) => {
-                const lonLat = toLonLat(coord);
-                return { lat: lonLat[1], lon: lonLat[0] };
-            });
-            const vertices = allVertices.filter((v, i) => {
-                if (i === 0) return true;
-                const prev = allVertices[i - 1];
-                return Math.abs(v.lat - prev.lat) > 1e-10 || Math.abs(v.lon - prev.lon) > 1e-10;
-            });
-
-            if (!this.dispatch || vertices.length < 3) return;
-
-            this.dispatch({
-                type: JaiaActions.ADD_EXCLUSION_ZONE,
-                exclusionZone: { vertices },
-            });
-        });
-
         return this.draw;
-    }
-
-    /**
-     * Clears the stored Draw interaction reference
-     *
-     * @returns {void}
-     */
-    clearDrawInteraction() {
-        this.draw = null;
     }
 
     /**
